pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-assistant"
version = "0.1.0"
description = "Research assistant with PDF processing and vector embeddings"
dependencies = [
    "accelerate==1.12.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.3",
    "cbor==1.0.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cuda-bindings==12.9.4",
    "cuda-pathfinder==1.2.2",
    "datasets==4.6.1",
    "dill==0.4.0",
    "filelock==3.20.0",
    "FlagEmbedding==1.3.5",
    "frozenlist==1.8.0",
    "fsspec==2025.12.0",
    "h11==0.16.0",
    "hf-xet==1.3.2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface_hub==0.36.2",
    "idna==3.11",
    "ijson==3.5.0",
    "inscriptis==2.7.1",
    "ir_datasets==0.5.11",
    "Jinja2==3.1.6",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.10",
    "langchain-core==1.2.14",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.0.10",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.9",
    "langsmith==0.7.6",
    "lxml==6.0.2",
    "lz4==4.4.5",
    "markdown-it-py==4.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib==3.9.2",
    "mdurl==0.1.2",
    "mpmath==1.2.1",
    "multiprocess==0.70.18",
    "networkx==3.4.2",
    "numpy==1.24.3",
    "nvidia-cublas-cu12",
    "nvidia-cuda-cupti-cu12",
    "nvidia-cuda-nvrtc-cu12",
    "nvidia-cuda-runtime-cu12",
    "nvidia-cudnn-cu12",
    "nvidia-cufft-cu12",
    "nvidia-curand-cu12",
    "nvidia-cusolver-cu12",
    "nvidia-cusparse-cu12",
    "nvidia-nccl-cu12",
    "nvidia-nvjitlink-cu12",
    "nvidia-nvtx-cu12",
    "packaging==24.2",
    "pandas==2.2.3",
    "peft==0.13.2",
    "pillow==11.1.0",
    "pinecone==6.0.0",
    "pydantic==2.7.4",
    "pydantic_core==2.18.4",
    "Pygments==2.18.0",
    "pymupdf==1.24.12",
    "pypdf==4.2.0",
    "pyright==1.1.386",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "PyYAML==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==13.8.1",
    "safetensors==0.4.5",
    "scipy==1.15.1",
    "sentencepiece==0.2.1",
    "setuptools==82.0.1",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sympy==1.13.1",
    "tabulate==0.9.0",
    "tokenizers==0.21.0",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "tqdm==4.67.1",
    "transformers==4.48.2",
    "typer==0.9.4",
    "typing_extensions==4.12.2",
    "urllib3==2.2.3",
    "xxhash==3.6.0",
    "zstandard==0.23.0",
]

[project.scripts]
respackage = "respackage.main:cli"

[tool.setuptools]
packages = ["app", "respackage"]
